backend/requirements.txt
asgiref==3.8.1
Django==4.2.3
django-cors-headers==4.4.0
djangorestframework==3.14.0
pillow==10.4.0
sqlparse==0.5.0
tzdata==2024.1
whitenoise==6.7.0